+++ pubtools/pulplib/_impl/model/repository/base.py
@@ -248,6 +248,42 @@ class Repository(PulpObject):
 
         return self._client._publish_repository(self, to_publish)
 
+    def remove_content(self, **kwargs):
+        """Remove all content of requested types from this repository.
+
+        Args:
+            type_ids (list[str])
+                IDs of content type(s) to be removed.
+                See :meth:`~pubtools.pulplib.Client.get_content_type_ids`.
+
+                If omitted, content of all types will be removed.
+
+        Returns:
+            Future[list[:class:`~pubtools.pulplib.Task`]]
+                A future which is resolved when content has been removed.
+
+                The future contains a list of zero or more tasks triggered and awaited
+                during the removal.
+
+                To obtain information on the removed content, use
+                :meth:`~pubtools.pulplib.Task.units`.
+
+        Raises:
+            DetachedException
+                If this instance is not attached to a Pulp client.
+
+        .. versionadded:: 1.5.0
+        """
+        if not self._client:
+            raise DetachedException()
+
+        # Note: we use dynamic kwargs because it's very likely that a future
+        # version of this method will support some "criteria".  Let's not fix the
+        # argument order at least until then.
+
+        type_ids = kwargs.get("type_ids")
+        return self._client._do_unassociate(self.id, type_ids)
+
     @classmethod
     def from_data(cls, data):
         # delegate to concrete subclass as needed